fastapi-api/requirements.txt
fastapi==0.110.0
numpy==1.26.4
pandas==2.2.1
pydantic==2.6.3
requests==2.31.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
SQLAlchemy==2.0.27
ujson==5.9.0
uvicorn==0.27.1
psycopg2==2.9.10